MATLAB window now supports "cleanup" functions
If a file named "foo_cleanup.m" exists in the same directory as "foo.m", and the MATLAB window executes function "foo" at least once, then function "foo_cleanup" will be executed when the MATLAB window resets (i.e. when the current experiment is unloaded).  It will be called with the last value returned by "foo" as its only argument.